+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@zodyac/zod-mongoose": "^1.1.3",
     "are-we-there-yet": "^4.0.2",
+    "axios": "^1.7.2",
     "bcrypt": "^5.1.1",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
