+++ src/routers/appRouter.js
@@ -4,6 +4,7 @@ import express from "express";
 import { raptorsRouter } from "./raptorsRouter.js";
 import { falconryRouter } from "./falconryRouter.js";
 import { userRouter } from "./userRouter.js";
+import { verifyEmailRouter } from "./verifyEmailRouter.js";
 
 export const appRouter = express.Router();
 
@@ -11,3 +12,5 @@ appRouter.use("/raptors", raptorsRouter);
 appRouter.use("/falconries", falconryRouter);
 
 appRouter.use("/user", userRouter); // by grouping all user-related routes under /user, I can centralize the handling of user data and actions. Also you can quickly identify where to add new user-related routes or modify existing ones.
+
+appRouter.use("/veri", verifyEmailRouter);
